Add files via upload
HTML files that include adding and editing a player, as well as the player list. Also included is the new database that was created, a singular table that takes and stores data for the players. Database module that works alongside the HTML files to help edit players. Simple UI to show everything.